+++ trilby-cli/src/Trilby/Install.hs
@@ -1,7 +1,6 @@
 module Trilby.Install (install) where
 
 import Data.Text qualified as Text
-import System.FilePath.Lens (directory)
 import Trilby.Disko (Disko)
 import Trilby.HNix (FlakeRef (..), currentSystem, writeNixFile)
 import Trilby.Install.Config.Host
@@ -11,17 +10,16 @@ import Trilby.Install.Disko qualified as Disko
 import Trilby.Install.Flake
 import Trilby.Install.Options
 import Trilby.Widgets
-import Turtle ((</>))
 import Prelude
 
-rootMount :: FilePath
-rootMount = "/mnt"
+rootMount :: Path Abs Dir
+rootMount = $(mkAbsDir "/mnt")
 
-trilbyDir :: FilePath
-trilbyDir = rootMount </> "etc/trilby"
+trilbyDir :: Path Abs Dir
+trilbyDir = trilbyHome rootMount
 
-diskoFile :: FilePath
-diskoFile = "/tmp/disko.nix"
+diskoFile :: Path Abs File
+diskoFile = $(mkAbsFile "/tmp/disko.nix")
 
 install :: InstallOpts Maybe -> App ()
 install (askOpts -> opts) | Just FlakeOpts{..} <- opts.flake = do
@@ -31,13 +29,13 @@ install (askOpts -> opts) | Just FlakeOpts{..} <- opts.flake = do
     nixosInstall flakeRef
     whenM copyFlake do
         storePath <- Text.strip <$> shell ("nix flake archive --json " <> flakeRef.url <> " | jq --raw-output .path") empty
-        asRoot cmd_ ["cp", "-r", storePath, fromString trilbyDir]
-        asRoot cmd_ ["chown", "-R", "1000:1000", fromString trilbyDir]
+        asRoot cmd_ ["cp", "-r", storePath, fromPath trilbyDir]
+        asRoot cmd_ ["chown", "-R", "1000:1000", fromPath trilbyDir]
     reboot opts.reboot
 install (askOpts -> opts) = do
     disko <- getDisko opts
-    inDir (diskoFile ^. directory) $ writeNixFile diskoFile disko
-    let diskoRef = Disko.File diskoFile
+    inDir (parent diskoFile) $ writeNixFile diskoFile disko
+    let diskoRef = Disko.File $ Abs diskoFile
     formatDisk opts.format diskoRef
     mountRoot diskoRef
     flakeRef <- setupHost disko opts
@@ -56,7 +54,12 @@ mountRoot d = unlessM rootIsMounted do
         errorExit "Cannot install without mounted partitions"
     disko $ Mount d
   where
-    rootIsMounted = (ExitSuccess ==) . fst <$> cmd' ["mountpoint", "-q", fromString rootMount]
+    rootIsMounted = (ExitSuccess ==) . fst <$> cmd' ["mountpoint", "-q", fromPath rootMount]
+
+flakeNix, defaultNix, configurationNix :: Path Rel File
+flakeNix = $(mkRelFile "flake.nix")
+defaultNix = $(mkRelFile "default.nix")
+configurationNix = $(mkRelFile "configuration.nix")
 
 setupHost :: Disko -> InstallOpts App -> App FlakeRef
 setupHost disko opts = do
@@ -64,24 +67,25 @@ setupHost disko opts = do
     edition <- opts.edition
     release <- opts.release
     inDir trilbyDir do
-        asRoot cmd_ ["chown", "-R", "1000:1000", fromString trilbyDir]
-        writeNixFile "flake.nix" $ flake release
+        asRoot cmd_ ["chown", "-R", "1000:1000", fromPath trilbyDir]
+        writeNixFile flakeNix $ flake release
         username <- opts.username
         password <- do
             rawPassword <- opts.password
             HashedPassword . firstLine <$> proc ["mkpasswd", rawPassword] empty
         let user = User{uid = 1000, ..}
-        let userDir = "users" </> fromText username
-        inDir userDir $ writeNixFile "default.nix" user
+        userDir <- parseRelDir . fromText $ "users/" <> username
+        let userFile = userDir </> defaultNix
+        writeNixFile userFile user
         keyboard <- opts.keyboard
         locale <- opts.locale
         timezone <- opts.timezone
         let host = Host{..}
-        let hostDir = "hosts" </> fromText hostname
+        hostDir <- parseRelDir . fromText $ "hosts/" <> hostname
         inDir hostDir do
             platform <- currentSystem
             writeNixFile
-                "default.nix"
+                defaultNix
                 [nix|
                 { lib, ... }:
                 lib.trilbySystem {
@@ -93,23 +97,31 @@ setupHost disko opts = do
                   modules = lib.findModulesList ./.;
                 }
                 |]
-            writeNixFile "configuration.nix" host
-            writeFile "hardware-configuration.nix"
+            writeNixFile configurationNix host
+            writeFile $(mkRelFile "hardware-configuration.nix")
                 =<< asRoot
                     cmd
                     [ "nixos-generate-config"
                     , "--show-hardware-config"
                     , "--no-filesystems"
                     , "--root"
-                    , fromString rootMount
+                    , fromPath rootMount
                     ]
-            writeNixFile "disko.nix" $ sanitise disko
+            writeNixFile $(mkRelFile "disko.nix") $ sanitise disko
         cmd_ . sconcat $
             [ ["nix", "flake", "lock"]
             , ["--accept-flake-config"]
             , ["--override-input", "trilby", "trilby"]
             ]
-    pure FlakeRef{url = fromString trilbyDir, output = pure hostname}
+        whenM opts.edit do
+            editor <- getEnv "EDITOR"
+            cmd_
+                [ fromString editor
+                , "flake.nix"
+                , fromPath $ hostDir </> configurationNix
+                , fromPath $ userDir </> defaultNix
+                ]
+    pure FlakeRef{url = fromPath trilbyDir, output = pure hostname}
 
 nixosInstall :: FlakeRef -> App ()
 nixosInstall flakeRef = do